move pdfjs dependent configs to it's package
