+++ src/siteConfig/pluginConfigs.ts
@@ -1,20 +1,7 @@
 import path from 'path';
 import dynamicRouterPlugin, { Config as DynamicRouteConfig } from '../plugins/plugin-dynamic-routes';
-import type { CurrentBundler, PluginConfig } from '@docusaurus/types';
-import CopyWebpackPlugin from 'copy-webpack-plugin';
+import type { PluginConfig } from '@docusaurus/types';
 import { sentryWebpackPlugin } from '@sentry/webpack-plugin';
-
-const pdfjsDistPath = path.dirname(require.resolve('pdfjs-dist/package.json'));
-
-const pdfjs_cMapsDir = path.join(pdfjsDistPath, 'cmaps');
-
-const getCopyPlugin = (currentBundler: CurrentBundler): typeof CopyWebpackPlugin => {
-    if (currentBundler.name === 'rspack') {
-        // @ts-expect-error: this exists only in Rspack
-        return currentBundler.instance.CopyRspackPlugin;
-    }
-    return CopyWebpackPlugin;
-};
 
 // TODO: Consider bundling default / recommended plugins.
 
@@ -113,32 +100,6 @@ export const sentryPluginConfig: PluginConfig = () => {
                         authToken: SENTRY_AUTH_TOKEN,
                         org: SENTRY_ORG,
                         project: SENTRY_PROJECT
-                    })
-                ]
-            };
-        }
-    };
-};
-
-export const pdfjsCopyDependenciesPluginConfig: PluginConfig = () => {
-    return {
-        name: 'pdfjs-copy-dependencies',
-        configureWebpack(config, isServer, { currentBundler }) {
-            const Plugin = getCopyPlugin(currentBundler);
-            return {
-                resolve: {
-                    alias: {
-                        canvas: false
-                    }
-                },
-                plugins: [
-                    new Plugin({
-                        patterns: [
-                            {
-                                from: pdfjs_cMapsDir,
-                                to: 'cmaps/'
-                            }
-                        ]
                     })
                 ]
             };
